+++ amazon/dynamodb/client/src/main/kotlin/org/http4k/connect/amazon/dynamodb/mapper/DynamoDbPage.kt
@@ -1,0 +1,7 @@
+package org.http4k.connect.amazon.dynamodb.mapper
+
+data class DynamoDbPage<Document: Any, HashKey: Any, SortKey: Any>(
+    val items: List<Document>,
+    val nextHashKey: HashKey?,
+    val nextSortKey: SortKey?
+)